Add login based on pyjwt

+++ notepad/app.py
@@ -1,16 +1,109 @@
+import jwt
 import psycopg2
 from flask import Flask, request
 
 DB = psycopg2.connect('postgresql://postgres:postgres@db/postgres')
 app = Flask('notepad')
-DB_PARODY = {}
 
 
-@app.route('/api/notepad', methods=['POST'])
+KEY_FOR_JWT = 'IncursoIncursoIncurso'
+
+
+def jwt_validation(user_jwt):
+    # Function to validate JWT
+    user_id = jwt.decode(user_jwt, KEY_FOR_JWT, algorithms=["RS256"])
+    user_id = user_id.split('.')
+    user_id = user_id[0]
+    cur = DB.cursor()
+
+    sql = """
+        SELECT id FROM users WHERE id = %s;
+    """
+    cur.execute(sql, (user_id,))
+    row = cur.fetchone()
+
+    if row[0] is None:
+        return True
+    else:
+        return False
+
+
+@app.route('/api/notepad/registration', methods=['POST'])
+def registration_of_a_new_user():
+    json_args = request.json
+    login = json_args['login']
+    password = json_args['password']
+
+    sql = """
+        SELECT date_of_creation FROM users WHERE login = %s;
+    """
+    cur = DB.cursor()
+    cur.execute(sql, (login,))
+    row = cur.fetchone()
+
+    if row is None:
+        sql = """
+            INSERT INTO users (login, password, )
+            VALUES (%s, %s)
+            ON CONFLICT (login)
+            DO NOTHING;
+        """
+        cur.execute(sql, (login, password))
+        DB.commit()
+        cur.close()
+        return {
+            'status': 'successefully registered',
+            'login': login,
+        }
+    else:
+        cur.close()
+        return {
+            'status': 'login is alredy taken',
+        }
+
+
+@app.route('/api/notepad/login', methods=['POST'])
+def login():
+    json_args = request.json
+    login = json_args['login']
+    password = json_args['password']
+
+    sql = """
+        SELECT id FROM users WHERE login = %s, password = %s;
+    """
+    cur = DB.cursor()
+    cur.execute(sql, (login, password,))
+    row = cur.fetchone()
+    cur.close()
+
+    if row is None:
+
+        return {
+            'status': 'wrong login or password',
+        }
+    else:
+
+        generated_jwt = jwt.encode(
+            {row[0]: password}, KEY_FOR_JWT, algorithm='HS256')
+
+        return {
+            'jwt': generated_jwt,
+        }
+
+
+@ app.route('/api/notepad', methods=['POST'])
 def create_or_update_notepad():
     json_args = request.json
     name = json_args['name']
     text = json_args['text']
+    jwt = json_args['jwt']
+
+    if jwt_validation(jwt) is False:
+        return {
+            'status': 'Forbidden'
+        }, 403
+
+    cur = DB.cursor()
 
     sql = """
         INSERT INTO notepad (name, text)
@@ -21,7 +114,7 @@ def create_or_update_notepad():
                 updated_at = now()
         RETURNING id;
     """
-    cur = DB.cursor()
+
     cur.execute(sql, (name, text))
     row_id = cur.fetchone()[0]
     sql = """
@@ -40,8 +133,16 @@ def create_or_update_notepad():
     }
 
 
-@app.route('/api/notepad/<notepad_id>', methods=['GET'])
+@ app.route('/api/notepad/<notepad_id>', methods=['GET'])
 def get_notepad(notepad_id):
+
+    json_args = request.json
+    jwt = json_args['jwt']
+
+    if jwt_validation(jwt) is False:
+        return {
+            'status': 'Forbidden'
+        }, 403
 
     sql = """
         SELECT id, name, text, updated_at FROM notepad WHERE id = %s
@@ -65,15 +166,32 @@ def get_notepad(notepad_id):
 
 
 def init_db():
+    cur = DB.cursor()
+
+    # sql = """ DROP TABLE users CASCADE"""
+
+    sql = """
+    CREATE TABLE IF NOT EXISTS users (
+        user_uuid SERIAL NOT NULL PRIMARY KEY,
+        login TEXT NOT NULL UNIQUE,
+        password TEXT NOT NULL,
+        date_of_creation TIMESTAMP DEFAULT current_timestamp
+    )
+    """
+
+    cur.execute(sql)
+
+    # sql = """ DROP TABLE notepad"""
+
     sql = """
     CREATE TABLE IF NOT EXISTS notepad (
         id SERIAL NOT NULL,
         name TEXT NOT NULL UNIQUE,
         text TEXT NOT NULL,
-        updated_at TIMESTAMP DEFAULT current_timestamp
+        updated_at TIMESTAMP DEFAULT current_timestamp,
+        author TEXT REFERENCES users (login)
     )
     """
-    cur = DB.cursor()
     cur.execute(sql)
     DB.commit()
     cur.close()
@@ -81,6 +199,14 @@ def init_db():
 
 @app.route('/api/notepad/delete/<notepad_id>', methods=['GET'])
 def delete_notepad(notepad_id):
+
+    json_args = request.json
+    jwt = json_args['jwt']
+
+    if jwt_validation(jwt) is False:
+        return {
+            'status': 'Forbidden'
+        }, 403
 
     sql = """
         DELETE FROM notepad WHERE id = %s
